refactor: decompose agent.rs into module directory (#515)
Splits the 1103-line agent.rs into focused submodules:
- agent/frontmatter.rs: AgentFrontmatter, AgentIsolation, and parsing logic
- agent/types.rs: AgentParams, AgentOutput, TokenUsage
- agent/execute.rs: AgentTool struct, Tool impl, and API helpers
- agent/tests.rs: all 21 tests (20 unit + 1 ignored integration)
- agent/mod.rs: re-exports preserving the public API surface

All external references (lib.rs, app_runtime.rs, worktree_isolation.rs,
agent_example.rs) continue to work unchanged via the same module paths.

Closes #515
